fix: ORDER API 수정

+++ src/main/java/com/team/saver/market/coupon/dto/CouponResponse.java
@@ -13,6 +13,6 @@ public class CouponResponse {
 
     private String couponDescription;
 
-    private double saleRate;
+    private int saleRate;
 
 }
